# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gerenciamento de clientes",
        "version": "0.1"
    },
    "paths": {
        "/api/customer": {
            "post": {
                "tags": [
                    "Customer"
                ],
                "summary": "Cadastra um novo cliente",
                "operationId": "e788a12c4fbd7937feab964d2e133089",
                "requestBody": {
                    "description": "Dados do cliente",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "amount"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Leide das Neves"
                                    },
                                    "amount": {
                                        "type": "numeric",
                                        "format": "float",
                                        "example": 99.98
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cliente cadastrado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Customer"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/customer/{customerId}/update-balance": {
            "put": {
                "tags": [
                    "Customer"
                ],
                "summary": "Atualiza o saldo do cliente",
                "operationId": "048f6cd07f214a814c27155b19aeb347",
                "parameters": [
                    {
                        "name": "customerId",
                        "in": "path",
                        "description": "ID do cliente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Dados para atualizar o saldo",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "amount"
                                ],
                                "properties": {
                                    "amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 599.9
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Depósito enviado para a fila de processamento",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/customer/{customerId}": {
            "get": {
                "tags": [
                    "Customer"
                ],
                "summary": "Obtém o saldo do cliente por ID",
                "operationId": "bb57c03d1aa2a41df20ab0bb00b2eebb",
                "parameters": [
                    {
                        "name": "customerId",
                        "in": "path",
                        "description": "ID do cliente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna o saldo do cliente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "amount": {
                                            "type": "number",
                                            "format": "float"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Cliente não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Cliente não encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Customer": {
                "title": "Customer",
                "description": "Customer model",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "amount": {
                        "type": "number",
                        "format": "float"
                    },
                    "before_amount": {
                        "type": "number",
                        "format": "float",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Customer",
            "description": "Operações relacionadas a clientes"
        }
    ]
}